Networks handled seperatly
